rename finResourceByType
