+++ karols-backend/timeslots/findResourceByType.js
